// src/Client/TreeMon.Mobile/src/app/models/screen.ts
// Copyright 2015, 2017 TreeMon.org.
// Licensed under CPAL 1.0,  See license.txt  or go to http://treemon.org/docs/license.txt  for full license details.

export class Screen {

    Caption: string;

    Command: string;

    Field: string;

    Value: string;

    Type: string;

    Operator: string;

    Junction: string;

    Order: number;

    Selected: boolean;
}